Make SecretKeyGenerator and LibCryptoRng package private
