+++ src/com/amazon/corretto/crypto/provider/LibCryptoRng.java
@@ -9,7 +9,7 @@ import java.security.SecureRandomSpi;
 /**
  * A simple wrapper around the linked LibCrypto's RAND_bytes() API.
  */
-public class LibCryptoRng extends SecureRandom {
+class LibCryptoRng extends SecureRandom {
     public static final String ALGORITHM_NAME = "LibCryptoRng";
     private static final long serialVersionUID = 1L;
     private static final int MAX_SINGLE_REQUEST = 8192;
